+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 
 This document keeps track of changes between releases of the library.
 
+master
+------
+
+* BC BREAK: Adding link between ClassMetadata and PropertyMetadata
+
 v0.3.0
 ------